+++ Client/Internal/Response.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.IO;
 using System.Net;
 using System.Net.Http;
@@ -50,9 +49,9 @@ namespace Pathoschild.Http.Client.Internal
         /// <summary>Asynchronously retrieve the response body as a list of deserialized models.</summary>
         /// <typeparam name="T">The response model to deserialize into.</typeparam>
         /// <exception cref="ApiException">An error occurred processing the response.</exception>
-        public Task<List<T>> AsList<T>()
+        public Task<T[]> AsList<T>()
         {
-            return this.As<List<T>>();
+            return this.As<T[]>();
         }
 
         /// <summary>Asynchronously retrieve the response body as an array of <see cref="byte"/>.</summary>